Allow -timeIncrement to be a float
The increment can now be fractional seconds. The printing in
ParseTimeControl() and SendTimeControl() is done with %g format, so that
for integer increment there is backward compatibility with the old level
command.

+++ backend.h
@@ -189,7 +189,7 @@ void AdjustClock P((Boolean which, int dir));
 void DisplayBothClocks P((void));
 void EditPositionMenuEvent P((ChessSquare selection, int x, int y));
 void DropMenuEvent P((ChessSquare selection, int x, int y));
-int ParseTimeControl P((char *tc, int ti, int mps));
+int ParseTimeControl P((char *tc, float ti, int mps));
 void EscapeExpand(char *p, char *q);
 void ProcessICSInitScript P((FILE * f));
 void EditCommentEvent P((void));
